feat(agent): enhance test code writer prompt.
