+++ packages/agent/src/orchestrate/test/structures/IAutoBeTestScenarioArtifacts.ts
@@ -1,0 +1,5 @@
+export interface IAutoBeTestScenarioArtifacts {
+  sdk: Record<string, string>;
+  dto: Record<string, string>;
+  e2e: Record<string, string>;
+}
